pyproject.toml
[tool.poetry]
name = "zillqa-crypto-analytics-etl"
version = "0.1.0"
description = ""
authors = ["codefox <isuranimalasiri@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.4"
google-cloud-bigquery = "^3.17.2"
boto3 = "^1.34.37"
awscli-local = "^0.22.0"
pyarrow = "^15.0.0"
pandas = "^2.2.0"
pyyaml = "^6.0.1"
localstack-client = "^2.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
